导入c端代码

+++ sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -1,17 +1,15 @@
 package com.sky.controller.admin;
 
-import com.github.pagehelper.Page;
 import com.sky.dto.DishDTO;
 import com.sky.dto.DishPageQueryDTO;
 import com.sky.entity.Dish;
 import com.sky.result.PageResult;
 import com.sky.result.Result;
-import com.sky.service.DishService;
+import com.sky.service.admin.DishService;
 import com.sky.vo.DishVO;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.collections4.Get;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -81,7 +79,7 @@ public class DishController {
         return Result.success();
     }
     @GetMapping("/list")
-    public Result<List<Dish>> findByCategoryId(Integer categoryId){
+    public Result<List<Dish>> findByCategoryId(Long categoryId){
         List<Dish> list = dishService.findByCategoryId(categoryId);
         return Result.success(list);
     }
